Extract constants from components
