+++ packages/zent/src/placeholder/shapes/consts.ts
@@ -1,0 +1,10 @@
+export const DEFAULT_SEGMENTS = [
+  [61.8, 38],
+  [30, 25, 44],
+  [20, 75],
+  [33, 16, 20, 27],
+  [12, 32, 53],
+  [45, 12, 42],
+  [20, 10, 47, 18],
+  [14, 47, 37],
+];
